refactor: createUserLotto 메소드 기능 분리, 수정
