+++ src/main/java/model/WinningLotto.java
@@ -1,16 +1,11 @@
 package model;
 
-import view.InputView;
-
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
-import static model.InputValidator.isValidateBonusNumber;
-import static model.InputValidator.isValidateWinningNumberLength;
+import static controller.LottoGameController.inputWinningNumber;
+import static view.InputView.inputBonusNumber;
 
 public class WinningLotto {
-    private static final int LOTTO_MIN_RANDOM_NUMBER = 1;
     private final Lotto lotto;
     private final int bonusNo;
 
@@ -31,23 +26,8 @@ public class WinningLotto {
         return winningLotto;
     }
 
-    // 지난 주 당첨 번호 입력
-    private static List<Integer> inputWinningNumber() {
-        String[] winningNumber;
-
-        do {
-            String lottoNumber = InputView.inputWinningNumber(); // 지난주 당첨번호 입력
-            winningNumber = splitWinningNumber(lottoNumber); // 당첨번호 , 기준으로 split
-        }
-        while (!(isValidateWinningNumberLength(winningNumber)));
-
-        List<Integer> winningNumbers = new ArrayList(Arrays.asList(winningNumber));
-
-        return winningNumbers;
-    }
-
     // 지난 주 당첨번호 String > String[] 스플릿
-    private static String[] splitWinningNumber(String number) {
+    public static String[] splitWinningNumber(String number) {
         return number.split(",");
     }
 
@@ -56,17 +36,5 @@ public class WinningLotto {
         boolean matchBonus = userLotto.isContainsBonusNumber(bonusNo);
 
         return Rank.valueOf(matchCount, matchBonus);
-    }
-
-    // 지난 주 보너스 번호 입력
-    private static int inputBonusNumber() {
-        int bonusNumber = LOTTO_MIN_RANDOM_NUMBER;
-
-        do {
-            bonusNumber = InputView.inputBonusNumber();
-            System.out.println("");
-        } while (!(isValidateBonusNumber(bonusNumber)));
-
-        return bonusNumber;
     }
 }
